Refactored text/font stuff into its own submodule.
