+++ src/context.rs
@@ -1,7 +1,6 @@
 
 
 use sdl2::{self, Sdl};
-use sdl2::render::Renderer;
 use sdl2::video::Window;
 
 use std::fmt;
